Always manage global version
Ensure that the global version of rbenv is always managed and defaults to
'system' even when `rbenv::global` isn't used.

Achieved by rejigging classes around, so that `rbenv::global::default` is
always included by `rbenv`. This is subsequently inherited and the file
contents overridden by `rbenv::global`.

This is sufficiently more complicated than I would like. But it's
essentially the only way to do it in Puppet.